expose RUMER api
