+++ server/controllers/stock/functions/README.md
@@ -1,0 +1,1 @@
+# This folder contains functions for API and reports
